Adds support for many Jobs in CPI manifest
The work to add CPI plugin support to the vSphere CPI has been
successful, but one substantial problem is that you can't deploy
anything other than the CPI release during a create-env operation.

This commit makes code changes that are _probably_ correct in order to
permit the instruction of the CPI to deploy multiple releases while
performing CLI-driving-the-CPI operations like create-env.

NOTE: No tests have been updated, so we expect a whole bunch of tests
      will fail. We're waiting to field-test these changes to ensure
      they behave as desired before bothering to rework the unit tests.

Signed-off-by: Kenneth Lakin <kenneth.lakin@broadcom.com>