Updates to handle CPI multiple releases
This commit updates the CPI inspection and CPI execution code to handle
the situation where the CPI is deployed with multiple releases. If one
wants to run CPI plugins during `create-env`, one must be able to handle
releases that aren't the CPI release, and also be able to discover WHICH
release is the source of the `bin/cpi` executable, so as to run the CPI.

Without this commit, we get the following error when we attempt to run
`create-env` with a multi-release CPI:

```
Installing CPI:
 Rendering and uploading Jobs:
   Too many jobs rendered... oops?
Exit code 1
```

Do NOTE that we've not updated any tests, so we expect a whole bunch of
tests to fail... but we've field-tested the CPI changes and the code
works as intended.

Signed-off-by: Kenneth Lakin <kenneth.lakin@broadcom.com>