Refactor CPI release validation
Gets rid of the `validator.go` file and moves the validation logic into
`CpiInstaller` struct. The validator was doing one thing only, so it's
probably simpler to move that validation. This also helps make the code
cleaner and easier to unit test.

Also fixes a bunch of type issues in other tests which just instantiated
this validator.

Signed-off-by: Kenneth Lakin <kenneth.lakin@broadcom.com>